requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
jinja2==3.1.4
python-dotenv==1.0.1
PyMuPDF==1.24.10
groq==0.11.0
httpx==0.27.2
